fix handling of context timeout and error in general
QueryContext was resulting in an infinite loop and unbounded memory
usage when the query timed out. It seems there was some strange
behavior going on in NewError and SQLGetDiagRec in particular. See
comments in error.go.

Also fixed QueryContext so it checks for context cancelation/timeout
before starting and so it rounds the timeout up to 1 second.